# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/test-task-cd
  description: Generated server url
paths:
  /api/publishers/{id}:
    put:
      tags:
      - publisher-controller
      summary: "Изменение издателя с номером {id}"
      operationId: updatePublisher
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherUpdateDto"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublisherUpdateDto"
    delete:
      tags:
      - publisher-controller
      summary: "Удаление издателя с номером {id}"
      operationId: deletePublisher
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/books/{id}:
    put:
      tags:
      - book-controller
      summary: "Изменение существующей книги с номером {id}"
      operationId: updateBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdateDto"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookUpdateDto"
    delete:
      tags:
      - book-controller
      summary: "Удаление книги с номером {id}"
      operationId: deletePublisher_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/authors/{id}:
    put:
      tags:
      - author-controller
      summary: "Изменение автора с номером {id}"
      operationId: updateAuthor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorUpdateDto"
        required: true
      responses:
        "202":
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthorUpdateDto"
    delete:
      tags:
      - author-controller
      summary: "Удаление автора с номером {id}"
      operationId: deleteAuthor
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/publishers:
    get:
      tags:
      - publisher-controller
      summary: "Чтение всех издателей. Фильтр по издателю, его сайту; пагинация."
      operationId: getAllPublishers
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: site
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublisherResponseDto"
    post:
      tags:
      - publisher-controller
      summary: Создание нового издателя.
      operationId: createPublisher
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublisherCreateDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublisherResponseDto"
  /api/books:
    get:
      tags:
      - book-controller
      summary: "Чтение всех книг. Фильтр по названию книги, автору, году издания,\
        \ издателю, bbk, isbn, количеству страниц. Пагинация."
      operationId: getAllBooks
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: author
        in: query
        required: false
        schema:
          type: string
      - name: printYearFrom
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: printYearTo
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: publisher
        in: query
        required: false
        schema:
          type: string
      - name: bbk
        in: query
        required: false
        schema:
          type: string
      - name: isbn
        in: query
        required: false
        schema:
          type: string
      - name: pagesFrom
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: pagesTo
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookResponseDto"
    post:
      tags:
      - book-controller
      summary: Создание новой книги.
      operationId: createBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookCreateDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookResponseDto"
  /api/authors:
    get:
      tags:
      - author-controller
      summary: "Чтение всех авторов. Фильтр по автору, пагинация"
      operationId: getAuthors
      parameters:
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuthorResponseDto"
    post:
      tags:
      - author-controller
      summary: Создание нового автора
      operationId: createAuthor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthorCreateDto"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthorResponseDto"
components:
  schemas:
    PublisherUpdateDto:
      type: object
      properties:
        name:
          type: string
        site:
          type: string
    BookUpdateDto:
      type: object
      properties:
        name:
          type: string
        authorIds:
          type: array
          items:
            type: integer
            format: int64
        printYear:
          type: integer
          format: int32
        publisherId:
          type: integer
          format: int64
        bbk:
          type: string
        isbn:
          type: string
        pages:
          type: integer
          format: int32
    AuthorUpdateDto:
      type: object
      properties:
        name:
          type: string
    PublisherCreateDto:
      type: object
      properties:
        name:
          type: string
        site:
          type: string
    PublisherResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        site:
          type: string
    BookCreateDto:
      type: object
      properties:
        name:
          type: string
        authorIds:
          type: array
          items:
            type: integer
            format: int64
        printYear:
          type: integer
          format: int32
        publisherId:
          type: integer
          format: int64
        bbk:
          type: string
        isbn:
          type: string
        pages:
          type: integer
          format: int32
    AuthorResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    BookResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        authors:
          type: array
          items:
            $ref: "#/components/schemas/AuthorResponseDto"
        printYear:
          type: integer
          format: int32
        publisher:
          $ref: "#/components/schemas/PublisherResponseDto"
        bbk:
          type: string
        isbn:
          type: string
        pages:
          type: integer
          format: int32
    AuthorCreateDto:
      type: object
      properties:
        name:
          type: string